+++ src/modules/human-resources/ifs-site.rv.ts
@@ -1,0 +1,17 @@
+import { isAuth } from '@/middlewares/is-auth';
+import { mapError } from '@/utils/map-error';
+import { Query, Resolver, UseMiddleware } from 'type-graphql';
+import { SiteTab } from './entities/ifs-site';
+
+@Resolver(SiteTab)
+export class IfsSiteResolver {
+  @Query(() => [SiteTab])
+  @UseMiddleware(isAuth)
+  async getAllSite(): Promise<SiteTab[] | undefined> {
+    try {
+      return await SiteTab.find({ order: { contract: 'ASC' } });
+    } catch (err) {
+      throw new Error(mapError(err));
+    }
+  }
+}